Crate aoc_bitvec

Dynamically sized array of items encoded as bits and packed

Modules

prelude

Structs

BitVec

Dynamically sized array of items encoded as bits and packed

BitVecEncodingDefault

Default implementation of BitVecEncoding for items which implement BitVecNative

BitVecIter

Iterator over the items in a BitVec

Traits

BitVecEncoding

Trait for encoding an item as bits for storing in a BitVec

BitVecNative

Trait for items which know how to encode themselves for storing in a BitVec